moving safety settings to vertex rest
